requirements.txt
aioredis==1.3.1
asgiref==3.3.1
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
certifi==2020.11.8
cffi==1.14.3
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
colorama==0.4.4
constantly==15.1.0
cryptography==3.2.1
daphne==2.5.0
dateparser==1.0.0
Django==3.1.5
django-cacheops==5.1
django-cleanup==5.1.0
django-crispy-forms==1.9.2
django-etc==1.2.0
django-hide==0.0.2
django-hitcount==1.3.3
django-multi-captcha-admin==1.0.0
django-ranged-response==0.2.0
django-recaptcha2==1.4.1
django-redis==4.12.1
django-translated-fields==0.10.2
django-utils-six==2.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
funcy==1.15
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==0.6.2
mysqlclient==2.0.1
Pillow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pylint==2.6.0
pyOpenSSL==19.1.0
python-binance==0.7.5
python-dateutil==2.8.1
python-decouple==3.3
python-gettext==4.0
pytz==2020.5
redis==3.5.3
regex==2020.10.28
requests==2.24.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
tzlocal==2.1
urllib3==1.25.11
wrapt==1.12.1
zope.interface==5.2.0
